requirements.txt
Python>3.7
Django>=2.1
Mysql>=5.7
Bootstrab4
django-phonenumber-field
